src/main/apache_airflow_app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.22.1
alembic==1.14.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
apache-airflow==2.10.3
apache-airflow-providers-common-compat==1.2.1
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.19.0
apache-airflow-providers-fab==1.5.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.2
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.7.1
apprise==1.9.6
argcomplete==3.5.1
asgi-lifespan==2.1.0
asgiref==3.8.1
asyncpg==0.31.0
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
beartype==0.22.9
billiard==4.2.4
blinker==1.8.2
cachelib==0.9.0
cachetools==6.2.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==14.0
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
coolname==2.2.0
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==42.0.8
dateparser==1.2.2
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.17.0
email_validator==2.2.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.127.1
filelock==3.20.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.10.0
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.3.0
griffe==1.15.0
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.6
httptools==0.7.1
httpx==0.27.0
humanfriendly==10.0
humanize==4.15.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.6.1
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lupa==2.6
Mako==1.3.6
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
multidict==6.1.0
numpy==2.4.0
oauthlib==3.3.1
omegaconf==2.3.0
ordered-set==4.1.0
orjson==3.11.5
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.5.1
pluggy==1.5.0
polars==1.36.1
polars-runtime-32==1.36.1
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.2.0
psutil==6.1.0
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyarrow==22.0.0
pycparser==2.22
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.3
Pygments==2.18.0
PyJWT==2.9.0
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.1
redis==7.1.0
referencing==0.35.1
regex==2025.11.3
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.20.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
semver==3.0.4
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
starlette==0.50.0
structlog==25.5.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
tomli==2.3.0
tomlkit==0.13.3
toposort==1.10
tqdm==4.67.1
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uc-micro-py==1.0.3
universal_pathlib==0.2.5
urllib3==2.2.3
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.2.1
yarl==1.17.1
zipp==3.20.2
